[#40991973] Automated Deployment Scripts
Deployment script now archives previous version of the site.

+++ web/config/scripts/knotisenv/deployment/site_copy.sh
@@ -5,6 +5,13 @@
 #                               #
 #===============================#
 
-rm -rf /srv/knotis/app/prev 2> /dev/null
-mv -f /srv/knotis/app/web /srv/knotis/app/prev
+# Backup old version of the site
+if [[ -d /srv/knotis/app/web ]] ; then
+    mkdir -p /srv/knotis/archive/ 2> /dev/null
+    tar -czf /srv/knotis/archive/knotis_"$(date +%Y_%m_%d_%H%M%S)".tar.gz /srv/knotis/app/web
+    rm -rf /srv/knotis/app/prev 2> /dev/null
+    mv -f /srv/knotis/app/web /srv/knotis/app/prev
+
+fi
+
 cp ${1} /srv/knotis/app/web
